+++ lenstronomy/Plots/output_plots.py
@@ -17,10 +17,11 @@ from lenstronomy.Data.imaging_data import ImageData
 def text_description(ax, d, text, color='w', backgroundcolor='k',
                      flipped=False, font_size=15):
     if flipped:
-        ax.text(d - d / 40., d - d / 15., text, color=color, fontsize=font_size,
+        ax.text(d - d / 40., d - d * font_size / 15.**2, text, color=color,
+                fontsize=font_size,
                 backgroundcolor=backgroundcolor)
     else:
-        ax.text(d / 40., d - d / 15., text, color=color, fontsize=font_size,
+        ax.text(d / 40., d - d * font_size / 15.**2, text, color=color, fontsize=font_size,
                 backgroundcolor=backgroundcolor)
 
 
@@ -450,6 +451,17 @@ class ModelPlot(object):
         plot_band = self._select_band(band_index)
         return plot_band.plot_subtract_from_data_all()
 
+    def plot_extinction_map(self, band_index=0, **kwargs):
+        """
+
+        :param band_index: index of band
+        :param kwargs: arguments of plotting
+        :return: plot instance of differential extinction map
+        """
+        plot_band = self._select_band(band_index)
+        return plot_band.plot_extinction_map(**kwargs)
+
+
     def source(self, band_index=0, **kwargs):
         """
 
@@ -472,7 +484,8 @@ class ModelBandPlot(object):
 
         self.bandmodel = SingleBandMultiModel(multi_band_list, kwargs_model,
                                                   likelihood_mask_list=likelihood_mask_list, band_index=band_index)
-        kwarks_lens_partial, kwargs_source_partial, kwargs_lens_light_partial, kwargs_ps_partial, kwargs_extinction_partial = self.bandmodel.select_kwargs(**kwargs_params)
+        self._kwargs_special_partial = kwargs_params.get('kwargs_special', None)
+        kwarks_lens_partial, kwargs_source_partial, kwargs_lens_light_partial, kwargs_ps_partial, self._kwargs_extinction_partial = self.bandmodel.select_kwargs(**kwargs_params)
         self._kwargs_lens_partial, self._kwargs_source_partial, self._kwargs_lens_light_partial, self._kwargs_ps_partial = self.bandmodel.update_linear_kwargs(param, kwarks_lens_partial, kwargs_source_partial, kwargs_lens_light_partial, kwargs_ps_partial)
         self._norm_residuals = self.bandmodel.reduced_residuals(model, error_map=error_map)
         self._reduced_x2 = self.bandmodel.reduced_chi2(model, error_map=error_map)
@@ -538,12 +551,13 @@ class ModelBandPlot(object):
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
 
-        scale_bar(ax, self._frame_size, dist=1, text='1"')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="w",
                          backgroundcolor='k', font_size=font_size)
 
         if 'no_arrow' not in kwargs or not kwargs['no_arrow']:
-            coordinate_arrows(ax, self._frame_size, self._coords, color='w', arrow_size=self._arrow_size)
+            coordinate_arrows(ax, self._frame_size, self._coords, color='w',
+                              arrow_size=self._arrow_size, font_size=font_size)
 
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
@@ -571,11 +585,13 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="w",
                          backgroundcolor='k', font_size=font_size)
         if 'no_arrow' not in kwargs or not kwargs['no_arrow']:
-            coordinate_arrows(ax, self._frame_size, self._coords, color='w', arrow_size=self._arrow_size)
+            coordinate_arrows(ax, self._frame_size, self._coords,
+                              color='w', arrow_size=self._arrow_size,
+                              font_size=font_size)
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
         cb = plt.colorbar(im, cax=cax)
@@ -609,10 +625,10 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"', color='w')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', color='w', font_size=font_size)
         if 'no_arrow' not in kwargs or not kwargs['no_arrow']:
             coordinate_arrows(ax, self._frame_size, self._coords, color='w',
-                              arrow_size=self._arrow_size)
+                              arrow_size=self._arrow_size, font_size=font_size)
         text_description(ax, self._frame_size, text=text,
                          color="w", backgroundcolor='k', flipped=False,
                          font_size=font_size)
@@ -641,11 +657,13 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k',
+                  font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="k",
                          backgroundcolor='w', font_size=font_size)
         if not no_arrow:
-            coordinate_arrows(ax, self._frame_size, self._coords, color='w', arrow_size=self._arrow_size)
+            coordinate_arrows(ax, self._frame_size, self._coords, color='w',
+                              arrow_size=self._arrow_size, font_size=font_size)
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
         cb = plt.colorbar(im, cax=cax)
@@ -667,10 +685,14 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k',
+                  font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="k",
                          backgroundcolor='w', font_size=font_size)
-        coordinate_arrows(ax, self._frame_size, self._coords, color='k', arrow_size=self._arrow_size)
+        coordinate_arrows(ax, self._frame_size, self._coords,
+                          font_size=font_size,
+                          color='k',
+                          arrow_size=self._arrow_size)
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
         cb = plt.colorbar(im, cax=cax)
@@ -710,6 +732,7 @@ class ModelBandPlot(object):
     def source_plot(self, ax, numPix, deltaPix_source, v_min=None,
                     v_max=None, with_caustics=False, caustic_color='yellow',
                     font_size=15, plot_scale='log',
+                    scale_size=0.1,
                     text="Reconstructed source",
                     colorbar_label=r'log$_{10}$ flux',
                     **kwargs):
@@ -734,6 +757,7 @@ class ModelBandPlot(object):
         d_s = numPix * deltaPix_source
         source, coords_source = self.source(numPix, deltaPix_source)
         if plot_scale == 'log':
+            source[source < 10**(v_min)] = 10**(v_min) # to remove weird shadow in plot
             source_scale = np.log10(source)
         elif plot_scale == 'linear':
             source_scale = source
@@ -755,15 +779,20 @@ class ModelBandPlot(object):
             ra_caustic_list, dec_caustic_list = self._caustics()
             plot_line_set(ax, coords_source, ra_caustic_list,
                           dec_caustic_list, color=caustic_color)
-        scale_bar(ax, d_s, dist=0.1, text='0.1"', color='w', flipped=False)
+        scale_bar(ax, d_s, dist=scale_size, text='{:.1f}"'.format(scale_size),
+                  color='w',
+                  flipped=False,
+                  font_size=font_size)
         if 'no_arrow' not in kwargs or not kwargs['no_arrow']:
-            coordinate_arrows(ax, self._frame_size, self._coords, color='w', arrow_size=self._arrow_size)
+            coordinate_arrows(ax, self._frame_size, self._coords, color='w',
+                              arrow_size=self._arrow_size, font_size=font_size)
         text_description(ax, d_s, text=text, color="w", backgroundcolor='k',
                          flipped=False, font_size=font_size)
         source_position_plot(ax, coords_source, self._kwargs_source_partial)
         return ax
 
-    def error_map_source_plot(self, ax, numPix, deltaPix_source, v_min=None, v_max=None, with_caustics=False):
+    def error_map_source_plot(self, ax, numPix, deltaPix_source, v_min=None,
+                              v_max=None, with_caustics=False, font_size=15):
         x_grid_source, y_grid_source = util.make_grid_transformed(numPix,
                                                                   self._coords.transform_pix2angle * deltaPix_source / self._deltaPix)
         x_center = self._kwargs_source_partial[0]['center_x']
@@ -783,13 +812,15 @@ class ModelBandPlot(object):
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
         cb = plt.colorbar(im, cax=cax)
-        cb.set_label(r'error variance', fontsize=15)
+        cb.set_label(r'error variance', fontsize=font_size)
         if with_caustics:
             ra_caustic_list, dec_caustic_list = self._caustics()
             plot_line_set(ax, coords_source, ra_caustic_list, dec_caustic_list, color='b')
-        scale_bar(ax, d_s, dist=0.1, text='0.1"', color='w', flipped=False)
-        coordinate_arrows(ax, d_s, coords_source, arrow_size=self._arrow_size, color='w')
-        text_description(ax, d_s, text="Error map in source", color="w", backgroundcolor='k', flipped=False)
+        scale_bar(ax, d_s, dist=0.1, text='0.1"', color='w', flipped=False, font_size=font_size)
+        coordinate_arrows(ax, d_s, coords_source,
+                          arrow_size=self._arrow_size, color='w', font_size=font_size)
+        text_description(ax, d_s, text="Error map in source", color="w",
+                         backgroundcolor='k', flipped=False, font_size=font_size)
         source_position_plot(ax, coords_source, self._kwargs_source_partial)
         return ax
 
@@ -817,9 +848,10 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k', font_size=font_size)
         if not no_arrow:
-            coordinate_arrows(ax, self._frame_size, self._coords, color='k', arrow_size=self._arrow_size)
+            coordinate_arrows(ax, self._frame_size, self._coords, color='k',
+                              arrow_size=self._arrow_size, font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="k",
                          backgroundcolor='w', font_size=font_size)
         divider = make_axes_locatable(ax)
@@ -854,8 +886,9 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k')
-        coordinate_arrows(ax, self._frame_size, self._coords, color='k', arrow_size=self._arrow_size)
+        scale_bar(ax, self._frame_size, dist=1, text='1"', color='k', font_size=font_size)
+        coordinate_arrows(ax, self._frame_size, self._coords, color='k',
+                          arrow_size=self._arrow_size, font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="k",
                          backgroundcolor='w', font_size=font_size)
         divider = make_axes_locatable(ax)
@@ -872,7 +905,10 @@ class ModelBandPlot(object):
         source_position_plot(ax, self._coords, self._kwargs_source_partial)
         return ax
 
-    def decomposition_plot(self, ax, text='Reconstructed', v_min=None, v_max=None, unconvolved=False, point_source_add=False, source_add=False, lens_light_add=False, **kwargs):
+    def decomposition_plot(self, ax, text='Reconstructed', v_min=None, v_max=None,
+                           unconvolved=False, point_source_add=False,
+                           font_size=15,
+                           source_add=False, lens_light_add=False, **kwargs):
         """
 
         :param ax:
@@ -900,16 +936,21 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"')
+        scale_bar(ax, self._frame_size, dist=1, text='1"', font_size=font_size)
         text_description(ax, self._frame_size, text=text, color="w", backgroundcolor='k')
-        coordinate_arrows(ax, self._frame_size, self._coords, arrow_size=self._arrow_size)
+        coordinate_arrows(ax, self._frame_size, self._coords,
+                          arrow_size=self._arrow_size, font_size=font_size)
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
         cb = plt.colorbar(im, cax=cax)
-        cb.set_label(r'log$_{10}$ flux', fontsize=15)
+        cb.set_label(r'log$_{10}$ flux', fontsize=font_size)
         return ax
 
-    def subtract_from_data_plot(self, ax, text='Subtracted', v_min=None, v_max=None, point_source_add=False, source_add=False, lens_light_add=False):
+    def subtract_from_data_plot(self, ax, text='Subtracted', v_min=None,
+                                v_max=None, point_source_add=False,
+                                source_add=False, lens_light_add=False,
+                                font_size=15
+                                ):
         model = self.bandmodel.image(self._kwargs_lens_partial, self._kwargs_source_partial, self._kwargs_lens_light_partial,
                                           self._kwargs_ps_partial, unconvolved=False, source_add=source_add,
                                           lens_light_add=lens_light_add, point_source_add=point_source_add)
@@ -922,13 +963,15 @@ class ModelBandPlot(object):
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.autoscale(False)
-        scale_bar(ax, self._frame_size, dist=1, text='1"')
-        text_description(ax, self._frame_size, text=text, color="w", backgroundcolor='k')
-        coordinate_arrows(ax, self._frame_size, self._coords, arrow_size=self._arrow_size)
+        scale_bar(ax, self._frame_size, dist=1, text='1"', font_size=font_size)
+        text_description(ax, self._frame_size, text=text, color="w",
+                         backgroundcolor='k', font_size=font_size)
+        coordinate_arrows(ax, self._frame_size, self._coords,
+                          arrow_size=self._arrow_size, font_size=font_size)
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
         cb = plt.colorbar(im, cax=cax)
-        cb.set_label(r'log$_{10}$ flux', fontsize=15)
+        cb.set_label(r'log$_{10}$ flux', fontsize=font_size)
         return ax
 
     def plot_main(self, with_caustics=False):
@@ -988,6 +1031,25 @@ class ModelBandPlot(object):
         f.tight_layout()
         f.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0., hspace=0.05)
         return f, axes
+
+    def plot_extinction_map(self, ax, v_min=None, v_max=None, **kwargs):
+        """
+
+        :param ax:
+        :param v_min:
+        :param v_max:
+        :return:
+        """
+        model = self.bandmodel.extinction_map(self._kwargs_extinction_partial, self._kwargs_special_partial)
+        if v_min is None:
+            v_min = 0
+        if v_max is None:
+            v_max = 1
+
+        im = ax.matshow(model, origin='lower', vmin=v_min, vmax=v_max,
+                        extent=[0, self._frame_size, 0, self._frame_size], **kwargs)
+        return ax
+
 
 
 def plot_chain_list(chain_list, index=0, num_average=100):